Refactored and TODO's Added
