+++ lib/ui/screens/Authentication/Register_Screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:help_me/main.dart';
-import 'package:help_me/shared/loading.dart';
 import 'package:help_me/services/firebase_auth.dart';
+import 'package:help_me/ui/shared/loading.dart';
 
 class RegisterScreen extends StatefulWidget {
   final Function toggleShowSignIn;
